requirements.txt
fastapi>=0.100,<0.101
uvicorn>=0.23.2,<0.24.0
chainlit==1.0.0
langchain-openai==0.1.0
httpx>=0.23.0,<0.25.0
python-dotenv==1.0.0
pydantic==2.5.0
gen3==4.24.1
langfuse 